requirements.txt
opencv-python>=4.8
numpy>=1.24
typer[all]>=0.9
rich>=13.7
scikit-image>=0.22
PyYAML>=6.0
pytest>=7.4
